# deepfence_server_client/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: community@deepfence.io
    name: Deepfence Support
    url: https://deepfence.io
  description: "Deepfence Runtime API provides programmatic control over Deepfence\
    \ microservice securing your container, kubernetes and cloud deployments. The\
    \ API abstracts away underlying infrastructure details like cloud provider, \n\
    container distros, container orchestrator and type of deployment. This is one\
    \ uniform API to manage and control security alerts, policies and response to\
    \ alerts for microservices running anywhere i.e. managed pure greenfield container\
    \ deployments or a mix of containers, VMs and serverless paradigms like AWS Fargate."
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  termsOfService: /tos
  title: Deepfence ThreatMapper
  version: 2.0.0
...
servers:
- url: /
...
paths:
  /deepfence/api-token:
    get:
      description: Get logged in user's API Tokens
      operationId: getApiTokens
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsBadRequestResponse'
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Internal Server Error
      security:
      - bearer_token: []
      summary: Get User's API Tokens
      tags:
      - User
...
  /deepfence/user/invite:
    post:
      description: Invite a user
      operationId: inviteUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelInviteUserRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsBadRequestResponse'
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Internal Server Error
      security:
      - bearer_token: []
      summary: Invite User
      tags:
      - User
  /deepfence/user/invite/register:
    post:
      description: Register invited user
      operationId: registerInvitedUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelRegisterInvitedUserRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsBadRequestResponse'
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Internal Server Error
      summary: Register Invited User
      tags:
      - User
...
  /deepfence/user/register:
    post:
      description: First user registration. Further users needs to be invited.
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelUserRegisterRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsBadRequestResponse'
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Internal Server Error
      summary: Register User
      tags:
      - User
  /deepfence/user/reset-password/request:
    post:
      description: Request for resetting the password
      operationId: resetPasswordRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelPasswordResetRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsBadRequestResponse'
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Internal Server Error
      summary: Reset Password Request
      tags:
      - User
...
components:
  schemas:
    ApiDocsBadRequestResponse:
      properties:
        error_fields:
          additionalProperties:
            type: string
          nullable: true
          type: object
        message:
          type: string
        success:
          example: false
          type: boolean
      type: object
    ApiDocsFailureResponse:
      properties:
        message:
          type: string
        success:
          example: false
          type: boolean
      type: object
...
    ModelInviteUserRequest:
      example:
        role: admin
        action: send-invite-email
        email: email
      properties:
        action:
          enum:
          - send-invite-email
          - get-invite-link
          type: string
        email:
          type: string
        role:
          enum:
          - admin
          - standard-user
          - read-only-user
          type: string
      required:
      - action
      - email
      - role
      type: object
...
    ModelPasswordResetRequest:
      example:
        email: email
      properties:
        email:
          type: string
      required:
      - email
      type: object
...
    ModelRegisterInvitedUserRequest:
      example:
        password: password
        code: code
        last_name: last_name
        first_name: first_name
        is_temporary_password: true
      properties:
        code:
          type: string
        first_name:
          type: string
        is_temporary_password:
          type: boolean
        last_name:
          type: string
        password:
          type: string
      required:
      - code
      - first_name
      - is_temporary_password
      - last_name
      - password
      type: object
    ModelResponse:
      example:
        data:
          access_token: access_token
          refresh_token: refresh_token
        success: true
        error_fields:
          key: error_fields
        message: message
      properties:
        data:
          $ref: '#/components/schemas/ModelResponseAccessToken'
        error_fields:
          additionalProperties:
            type: string
          nullable: true
          type: object
        message:
          type: string
        success:
          type: boolean
      type: object
    ModelResponseAccessToken:
      example:
        access_token: access_token
        refresh_token: refresh_token
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
      required:
      - access_token
      - refresh_token
      type: object
...
    ModelUserRegisterRequest:
      example:
        password: password
        console_url: console_url
        last_name: last_name
        company: company
        first_name: first_name
        is_temporary_password: true
        email: email
      properties:
        company:
          type: string
        console_url:
          type: string
        email:
          type: string
        first_name:
          type: string
        is_temporary_password:
          type: boolean
        last_name:
          type: string
        password:
          type: string
      required:
      - company
      - console_url
      - email
      - first_name
      - is_temporary_password
      - last_name
      - password
      type: object
...
  securitySchemes:
    bearer_token:
      bearerFormat: JWT
      description: RW Access
      scheme: bearer
      type: http