# deepfence_server_client/model_model_password_reset_request.go
/*
Deepfence ThreatMapper

Deepfence Runtime API provides programmatic control over Deepfence microservice securing your container, kubernetes and cloud deployments. The API abstracts away underlying infrastructure details like cloud provider,  container distros, container orchestrator and type of deployment. This is one uniform API to manage and control security alerts, policies and response to alerts for microservices running anywhere i.e. managed pure greenfield container deployments or a mix of containers, VMs and serverless paradigms like AWS Fargate.

API version: 2.0.0
Contact: community@deepfence.io
*/

// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

package deepfence_server_client

import (
	"encoding/json"
)

// checks if the ModelPasswordResetRequest type satisfies the MappedNullable interface at compile time
var _ MappedNullable = &ModelPasswordResetRequest{}

// ModelPasswordResetRequest struct for ModelPasswordResetRequest
type ModelPasswordResetRequest struct {
	Email string `json:"email"`
}

// NewModelPasswordResetRequest instantiates a new ModelPasswordResetRequest object
// This constructor will assign default values to properties that have it defined,
// and makes sure properties required by API are set, but the set of arguments
// will change when the set of required properties is changed
func NewModelPasswordResetRequest(email string) *ModelPasswordResetRequest {
	this := ModelPasswordResetRequest{}
	this.Email = email
	return &this
}

// NewModelPasswordResetRequestWithDefaults instantiates a new ModelPasswordResetRequest object
// This constructor will only assign default values to properties that have it defined,
// but it doesn't guarantee that properties required by API are set
func NewModelPasswordResetRequestWithDefaults() *ModelPasswordResetRequest {
	this := ModelPasswordResetRequest{}
	return &this
}

// GetEmail returns the Email field value
func (o *ModelPasswordResetRequest) GetEmail() string {
	if o == nil {
		var ret string
		return ret
	}

	return o.Email
}

// GetEmailOk returns a tuple with the Email field value
// and a boolean to check if the value has been set.
func (o *ModelPasswordResetRequest) GetEmailOk() (*string, bool) {
	if o == nil {
		return nil, false
	}
	return &o.Email, true
}

// SetEmail sets field value
func (o *ModelPasswordResetRequest) SetEmail(v string) {
	o.Email = v
}

func (o ModelPasswordResetRequest) MarshalJSON() ([]byte, error) {
	toSerialize,err := o.ToMap()
	if err != nil {
		return []byte{}, err
	}
	return json.Marshal(toSerialize)
}

func (o ModelPasswordResetRequest) ToMap() (map[string]interface{}, error) {
	toSerialize := map[string]interface{}{}
	toSerialize["email"] = o.Email
	return toSerialize, nil
}

type NullableModelPasswordResetRequest struct {
	value *ModelPasswordResetRequest
	isSet bool
}

func (v NullableModelPasswordResetRequest) Get() *ModelPasswordResetRequest {
	return v.value
}

func (v *NullableModelPasswordResetRequest) Set(val *ModelPasswordResetRequest) {
	v.value = val
	v.isSet = true
}

func (v NullableModelPasswordResetRequest) IsSet() bool {
	return v.isSet
}

func (v *NullableModelPasswordResetRequest) Unset() {
	v.value = nil
	v.isSet = false
}

func NewNullableModelPasswordResetRequest(val *ModelPasswordResetRequest) *NullableModelPasswordResetRequest {
	return &NullableModelPasswordResetRequest{value: val, isSet: true}
}

func (v NullableModelPasswordResetRequest) MarshalJSON() ([]byte, error) {
	return json.Marshal(v.value)
}

func (v *NullableModelPasswordResetRequest) UnmarshalJSON(src []byte) error {
	v.isSet = true
	return json.Unmarshal(src, &v.value)
}